2d/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pillow==10.2.0
PyOpenGL==3.1.7
pyopengltk==0.0.4
tk==0.1.0
typing_extensions==4.10.0